# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "WPPConnect API Rest",
    "description": "Welcome to the wppconnect-server API documentation. This API provides a set of endpoints to interact with the wppconnect-server application, allowing you to build integrations and automate interactions with WhatsApp."
  },
...
  "paths": {
...
    "/api/{session}/newsletter": {
      "post": {
        "tags": [
          "Newsletter"
        ],
        "description": "",
        "operationId": "createNewsletter",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "options": {
                    "type": "object"
                  }
                }
              },
              "examples": {
                "Create newsletter/channel": {
                  "value": {
                    "name": "Name for your channel",
                    "options": {
                      "description": "Description of channel",
                      "picture": "<base64_image>"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/{session}/newsletter/{id}": {
      "put": {
        "tags": [
          "Newsletter"
        ],
        "description": "",
        "operationId": "editNewsletter",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "<newsletter_id>"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "picture": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "Edit newsletter/channel": {
                  "value": {
                    "name": "New name of channel",
                    "description": "New description of channel",
                    "picture": "<new_base64_image> or send null"
                  }
                },
                "Create newsletter/channel": {
                  "value": {
                    "name": "Name for your channel",
                    "options": {
                      "description": "Description of channel",
                      "picture": "<base64_image>"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Newsletter"
        ],
        "description": "",
        "operationId": "destroyNewsletter",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NEWSLETTER ID"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/{session}/mute-newsletter/{id}": {
      "post": {
        "tags": [
          "Newsletter"
        ],
        "description": "",
        "operationId": "muteNewsletter",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NEWSLETTER ID"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
  },
...
}